#16 : Configurations CrossOrigin with front-End

+++ catalogue-microservice/src/main/java/org/cataloguemicroservice/dtos/ProductDTO.java
@@ -3,18 +3,21 @@ package org.cataloguemicroservice.dtos;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.index.Indexed;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductDTO {
-    private Integer productId;
+
+    private Long productId;
     private String label;
     private String slug;
     private String productTitle;
     private String imageUrl;
     private Long price;
     private String description;
-    private Integer idCategory;
+    private Long idCategory;
     private boolean valid;
 }
